Reworking the QDS architecture so it doesn't require the application developer to delegate QApplication creation to QDS.
The main reason for this requirement has been the desire to use a KDE
service implementation, which needs a KApplication instance.

The new architecture uses an out-of-process approach for implementing the
desktop API access with D-BUS (http://www.freedesktop.org/Software/dbus)
being used as the commmunication layer.

The new QDS-Daemon provides the services based on the same desktop wrapper
plugins formerly used directly by the QDS library, but the D-BUS approach
also opens the door for implemeting the service directly in the
respective desktop environemnt.